feat(ci): Renaming the modules prevents the local package from shadowing the external kafka library, so imports like KafkaProducer resolve correctly.
